Changed the implementation to use Buffer instead of the native file Copy, Also did some reorganization
